Release 1.1 - Echoing Sysloggian Bane
Added a noecho option to text_input which disables terminal echo and
refrains from Term::ReadLine magic.

Added pwprompt() to prompt a user for a password.
Modified prompt() to detect /passw(or)?d/i in a prompt and switch
automatically to pwprompt.

Added a 'no_syslog' option to output to replace '_skip_syslog' which was
a hidden feature. This allows a user to ensure certain messages never go
to syslog.

Updated documentation to reflect the current state of affairs.