Change default behavior if importing outside of main
It doesn't make much sense for a module you're including to mess with
`@ARGV` by default.  This patch uses `Hook::LexWrap` to wrap `import`
and check the calling package.  If the calling package is `main`, we'll
do the `@ARGV` parsing by default. If it's not, we'll skip the
initialization until it's used.  You can override this behavior if you
specify:

    package YOLO;

    use CLI::Helpers qw(:output at_import);

In the module.