Remove Sub::Exporter and handle import locally
Sub::Exporter is awesome, but I needed to provide a custom `import()`
function to be able to detect the importing package and change the
default behavior.  I had attempted this with `Hook::LexWrap`, but that
caused unexpected issues with the `import()` routine in certain
situations.  This code is cleaner, has fewer deps, but lacks the
extended functionality of `Sub::Exporter`.